--- packages/template-retail-react-app/app/components/apple-pay-express/utils/api.js
@@ -1,45 +1,0 @@
-/*
- * Copyright (c) 2025, Salesforce, Inc.
- * All rights reserved.
- * SPDX-License-Identifier: BSD-3-Clause
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-export class ApiClient {
-    url = null
-    token = null
-    site = null
-
-    constructor(url, token, site) {
-        this.url = url
-        this.token = token
-        this.site = site
-    }
-
-    base(method, options) {
-        const queryParams = {
-            siteId: this.site.id,
-            ...(options?.queryParams || {})
-        }
-        const fullUrl = `${this.url}?${new URLSearchParams(queryParams)}`
-
-        const requestConfig = {
-            method: method,
-            body: options?.body || null,
-            headers: {
-                'Content-Type': 'application/json',
-                authorization: `Bearer ${this.token}`,
-                ...options?.headers
-            }
-        }
-
-        return fetch(fullUrl, requestConfig)
-    }
-
-    get(options) {
-        return this.base('get', options)
-    }
-
-    post(options) {
-        return this.base('post', options)
-    }
-}
