+++ packages/template-retail-react-app/app/components/express/utils/api.test.js
@@ -6,15 +6,33 @@
  */
 import {ApiClient} from '@salesforce/retail-react-app/app/components/express/utils/api'
 
+// Mock fetch properly
+const mockFetch = jest.fn()
+global.fetch = mockFetch
+
+// Suppress MSW 'Found an unhandled' warnings for this test file
+const originalConsoleError = console.error
+beforeAll(() => {
+    console.error = (...args) => {
+        if (typeof args[0] === 'string' && args[0].startsWith('Found an unhandled')) {
+            return
+        }
+        originalConsoleError(...args)
+    }
+})
+afterAll(() => {
+    console.error = originalConsoleError
+})
+
 describe('ApiClient', () => {
     let apiClient
-    const mockUrl = '/api/test'
     const mockToken = 'test-token'
     const mockSite = {id: 'test-site'}
+    const mockUrl = '/api/test'
 
     beforeEach(() => {
         apiClient = new ApiClient(mockUrl, mockToken, mockSite)
-        jest.clearAllMocks()
+        mockFetch.mockClear()
     })
 
     describe('constructor', () => {
@@ -27,58 +45,86 @@ describe('ApiClient', () => {
 
     describe('base method', () => {
         it('should make a GET request with correct parameters', async () => {
-            const mockResponse = {status: 200, json: () => Promise.resolve({data: 'test'})}
-            const mockFetch = jest.fn().mockResolvedValue(mockResponse)
-            global.fetch = mockFetch
+            const mockResponse = {
+                ok: true,
+                json: () => Promise.resolve({}),
+                text: () => Promise.resolve(''),
+                clone: () => mockResponse,
+                headers: new Headers()
+            }
+            mockFetch.mockResolvedValue(mockResponse)
 
             await apiClient.base('GET', {
                 queryParams: {param1: 'value1'}
             })
 
-            expect(mockFetch).toHaveBeenCalledWith(
-                `${mockUrl}?siteId=${mockSite.id}&param1=value1`,
-                {
-                    method: 'GET',
-                    body: null,
-                    headers: {
-                        'Content-Type': 'application/json',
-                        authorization: `Bearer ${mockToken}`
-                    }
-                }
-            )
+            const request = mockFetch.mock.calls[0][0]
+            expect(request.url).toBe(`${mockUrl}?siteId=${mockSite.id}&param1=value1`)
+            expect(request.method).toBe('GET')
+            expect(request.headers.get('Content-Type')).toBe('application/json')
+            expect(request.headers.get('authorization')).toBe(`Bearer ${mockToken}`)
         })
 
         it('should make a POST request with body and headers', async () => {
-            const mockResponse = {status: 200, json: () => Promise.resolve({data: 'test'})}
-            const mockFetch = jest.fn().mockResolvedValue(mockResponse)
-            global.fetch = mockFetch
-
+            const mockResponse = {
+                ok: true,
+                json: () => Promise.resolve({}),
+                text: () => Promise.resolve(''),
+                clone: () => mockResponse,
+                headers: new Headers()
+            }
+            mockFetch.mockResolvedValue(mockResponse)
             const body = JSON.stringify({test: 'data'})
-            const headers = {custom: 'header'}
+            const customHeaders = {custom: 'header'}
 
             await apiClient.base('POST', {
                 body,
-                headers
+                headers: customHeaders
             })
 
-            expect(mockFetch).toHaveBeenCalledWith(`${mockUrl}?siteId=${mockSite.id}`, {
-                method: 'POST',
-                body,
-                headers: {
-                    'Content-Type': 'application/json',
-                    authorization: `Bearer ${mockToken}`,
-                    custom: 'header'
-                }
-            })
+            const request = mockFetch.mock.calls[0][0]
+            expect(request.url).toBe(`${mockUrl}?siteId=${mockSite.id}`)
+            expect(request.method).toBe('POST')
+            expect(request.headers.get('Content-Type')).toBe('application/json')
+            expect(request.headers.get('authorization')).toBe(`Bearer ${mockToken}`)
+            expect(request.headers.get('custom')).toBe('header')
+
+            // Check body separately to avoid conditional expect
+            const requestBody = await request.text()
+            expect(requestBody).toBe(body)
+        })
+
+        it('should handle request without optional parameters', async () => {
+            const mockResponse = {
+                ok: true,
+                json: () => Promise.resolve({}),
+                text: () => Promise.resolve(''),
+                clone: () => mockResponse,
+                headers: new Headers()
+            }
+            mockFetch.mockResolvedValue(mockResponse)
+
+            await apiClient.base('GET')
+
+            const request = mockFetch.mock.calls[0][0]
+            expect(request.url).toBe(`${mockUrl}?siteId=${mockSite.id}`)
+            expect(request.method).toBe('GET')
+            expect(request.body === null || request.body === undefined).toBe(true)
+            expect(request.headers.get('Content-Type')).toBe('application/json')
+            expect(request.headers.get('authorization')).toBe(`Bearer ${mockToken}`)
         })
     })
 
     describe('get method', () => {
         it('should call base with GET method', async () => {
-            const mockResponse = {status: 200, json: () => Promise.resolve({data: 'test'})}
-            const mockFetch = jest.fn().mockResolvedValue(mockResponse)
-            global.fetch = mockFetch
-
+            const mockResponse = {
+                ok: true,
+                json: () => Promise.resolve({}),
+                text: () => Promise.resolve(''),
+                clone: () => mockResponse,
+                headers: new Headers()
+            }
+            mockFetch.mockResolvedValue(mockResponse)
             const spy = jest.spyOn(apiClient, 'base')
 
             await apiClient.get({param: 'value'})
@@ -89,10 +135,14 @@ describe('ApiClient', () => {
 
     describe('post method', () => {
         it('should call base with POST method', async () => {
-            const mockResponse = {status: 200, json: () => Promise.resolve({data: 'test'})}
-            const mockFetch = jest.fn().mockResolvedValue(mockResponse)
-            global.fetch = mockFetch
-
+            const mockResponse = {
+                ok: true,
+                json: () => Promise.resolve({}),
+                text: () => Promise.resolve(''),
+                clone: () => mockResponse,
+                headers: new Headers()
+            }
+            mockFetch.mockResolvedValue(mockResponse)
             const spy = jest.spyOn(apiClient, 'base')
 
             await apiClient.post({param: 'value'})
