+++ packages/template-retail-react-app/app/components/express/utils/payment-methods.js
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import {ApiClient} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/api'
+import {ApiClient} from '@salesforce/retail-react-app/app/components/express/utils/api'
 
 export class AdyenPaymentMethodsService {
     baseUrl = '/api/adyen/paymentMethods/standalone'
