+++ packages/template-retail-react-app/app/components/express/utils/payment-methods.test.js
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import {AdyenPaymentMethodsService} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/payment-methods'
-import {ApiClient} from '@salesforce/retail-react-app/app/components/apple-pay-express/utils/api'
+import {AdyenPaymentMethodsService} from '@salesforce/retail-react-app/app/components/express/utils/payment-methods'
+import {ApiClient} from '@salesforce/retail-react-app/app/components/express/utils/api'
 
 // Mock the ApiClient
-jest.mock('@salesforce/retail-react-app/app/components/apple-pay-express/utils/api')
+jest.mock('@salesforce/retail-react-app/app/components/express/utils/api')
 
 describe('AdyenPaymentMethodsService', () => {
     let paymentMethodsService
