+++ packages/template-retail-react-app/app/components/express/utils/pdp/basket-calculation.js
@@ -156,4 +156,4 @@ export const forceOrderCalculation = async (basketId, authToken, site) => {
     } catch (error) {
         throw error
     }
-}
+}
